requirements.txt
pandas>=2.0.0
numpy>=1.26.0
openpyxl>=3.1.0
SQLAlchemy>=2.0.0
pyodbc>=5.0.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=5.0.0